% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the StatML Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
Run the StatML Shiny Application
}
