% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatML-package.R
\docType{package}
\name{StatML-package}
\alias{StatML}
\alias{StatML-package}
\title{StatML: Interactive Machine Learning Platform for Ecology and Data Science}
\description{
An interactive Shiny application for machine learning, part of the StatSuite ecosystem. Covers regression and classification using linear models, generalized linear models, and random forests with tidymodels. Designed for teaching and research in ecology, biodiversity sciences, and data science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuelSpinola/StatML}
  \item Report bugs at \url{https://github.com/ManuelSpinola/StatML/issues}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{manuel.spinola@una.cr}

Authors:
\itemize{
  \item Manuel Spínola \email{manuel.spinola@una.cr}
}

}
\keyword{internal}
